Enumクラスをリファクタリング
