+++ src/main/java/com/c4c/_2022server/enums/EnumBase.java
@@ -1,0 +1,6 @@
+package com.c4c._2022server.enums;
+
+interface EnumBase {
+    Integer getCode();
+    String getName();
+}
